requirements.txt
discord.py>=2.0.0
python-dotenv
yt_dlp
aiohttp
qrcode
python-barcode
Pillow
